translators.com.mk
A final exam project for a real life client, who is a translator non-profit organization. The project was assigned to all the students at the academy to firstly make a prototype design in XD, than upgraded into a real time website. This project was selected by the clients to be worked upon further more.